Move convertDateTime, refactor query_utils (#1064)
